chore: restructure Toolbar under ItemList feature for clearer module separation
