+++ client/src/components/layout/Toolbar/Toolbar.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import type { GenreKey } from "@/types/genre";
-import SearchBar from "@/features/Toolbar/atoms/SerchBar/SearchBar";
+import SearchBar from "@/components/layout/Toolbar/atoms/SerchBar/SearchBar";
 import Button from "@/components/atoms/Button/Button";
-import CategorySelector from "@/features/Toolbar/atoms/CategorySelector/CategorySelector";
+import CategorySelector from "@/components/layout/Toolbar/atoms/CategorySelector/CategorySelector";
 
 const Toolbar = () => {
 	//typeに関してdefault準備する必要もある
